Seggregated the TagCreatorBase tests
